fix: Resolve SonarQube security vulnerabilities and code smells
- Security fixes:
  * Remove hardcoded Twilio API credentials (planning-service)
  * Remove hardcoded Stripe webhook secrets (event-service)
  * Reduce file upload size limit from 50MB to 8MB (course-service)
  * Remove hardcoded passwords from PasswordHashGenerator
  * Add .env.example for environment variable documentation

- Bug fixes:
  * Fix long cast issue in EventAnalyticsService
  * Fix unused return value in ApplicationService
  * Add lang and xml:lang attributes to email templates (6 files)

- Code quality improvements:
  * Replace System.out with SLF4J logger in PasswordHashGenerator
  * Create custom exceptions (UserNotFoundException, DuplicateEmailException, AuthenticationException)
  * Use custom exceptions in AuthService instead of generic RuntimeException
  * Fix primitive boolean expressions in AuthService

- Refactoring:
  * Create constants classes to avoid string duplication:
    - RoleConstants (certification-badge, course, exam services)
    - LogConstants (certification-badge-service)
    - ErrorMessages (certification-badge, course, event services)
    - SuccessMessages (exam-service)
    - ApiEndpoints (course, exam services)
    - WebSocketConstants (course-service)
  * Update AuthorizationService to use RoleConstants
  * Update SignalingWebSocketHandler to use WebSocketConstants

This commit addresses multiple SonarQube issues including:
- 6 security vulnerabilities (Blocker/Major)
- 9 bugs
- 15+ critical code smells

+++ Backend/event-service/src/main/java/com/smartek/event/service/EventAnalyticsService.java
@@ -46,7 +46,7 @@ public class EventAnalyticsService {
                 (double) analytics.getConfirmedRegistrations() / event.getTotalCapacity() * 100 : 0;
         analytics.setFillRate(Math.round(fillRate * 100.0) / 100.0);
 
-        long totalRegistrations = analytics.getConfirmedRegistrations() + waitingListSize;
+        long totalRegistrations = (long) analytics.getConfirmedRegistrations() + waitingListSize;
         long cancelledCount = registrationRepository.countCancelledByEventId(eventId);
         double cancellationRate = totalRegistrations > 0 ? 
                 (double) cancelledCount / (totalRegistrations + cancelledCount) * 100 : 0;
